Move Systrace.h into android/Systrace.h
